moved contents into page file
