+++ pytest.ini
@@ -10,8 +10,8 @@ filterwarnings =
     ignore::RuntimeWarning
     ignore::sqlalchemy.exc.SAWarning
     ignore:.*coroutine.*was never awaited:RuntimeWarning
-    ignore:.*Exception terminating connection:sqlalchemy.pool
-    ignore:.*Future exception was never retrieved:asyncio
-    ignore:.*got result for unknown protocol state:asyncpg
-    ignore:.*Event loop is closed:RuntimeError
-    ignore:.*got Future.*attached to a different loop:RuntimeError
+    ignore:.*Exception terminating connection:RuntimeWarning
+    ignore:.*Future exception was never retrieved:RuntimeWarning
+    ignore:.*got result for unknown protocol state:RuntimeWarning
+    ignore:.*Event loop is closed:RuntimeWarning
+    ignore:.*got Future.*attached to a different loop:RuntimeWarning
